changed PluginManager to not be a gtk.Dialog, instead it creates one when needed, this allows the dialog to be recreated after receiving a destroy event
fixed bug where user could open multiple dialogs

removed debug messages

Unload Python plugins when GeanyPy is deactivated

This prevents crashes and/or lockups when the GeanyPy C plugin
is reloaded.

FIXME: Because it calls the normal unloading path, it updates
the list of loaded plugins to remove each plugin. The net effect
is that when the C GeanyPy plugin is unloaded, it forgets all
of the Python plugins that were loaded at the time.

Closes #24

changed PluginManager to not be a gtk.Dialog, instead it creates one when needed, this allows the dialog to be recreated after receiving a destroy event

fixed bug where user could open multiple dialogs

removed debug messages

changed PluginManager to not be a gtk.Dialog, instead it creates one when needed, this allows the dialog to be recreated after receiving a destroy event

fixed bug where user could open multiple dialogs

removed debug messages

added fix for python plugin manager dialog not getting destroyed when geanypy was unloaded

Conflicts:
	geany/manager.py