Update bundled ax_python_devel.m4 to latest upstream
This actually fixes building on Debian Unstable that ships
ax_python_devel.m4 in autoconf-archive, and apparently aclocal prefers
system directories.  So we get the system version, which happens to be
the latest, which changed API.

To avoid breaking older systems that would also ship ax_python_devel.m4
but in an older version with the older API, keep using the old API too.